totalPrice四舍五入

+++ src/main/java/com/youzan/sz/common/util/PriceUtil.java
@@ -30,7 +30,7 @@ public class PriceUtil {
      * @param amount
      * @return
      */
-    public static BigDecimal totalPrice(Long price, Long amount) {
+    public static BigDecimal totalPrice(Long price, Integer amount) {
         return new BigDecimal(price * amount).divide(new BigDecimal(Common.QUANTITY_MULTIPLE));
     }
 
@@ -41,14 +41,14 @@ public class PriceUtil {
      * @param amount
      * @return
      */
-    public static Long totalRoundingPrice(Long price, Long amount) {
+    public static Long totalRoundingPrice(Long price, Integer amount) {
         return  totalPrice(price, amount).setScale(0,BigDecimal.ROUND_HALF_UP).longValue();
     }
 
     public static void main(String[] args) {
 //        System.out.println(new BigDecimal(0.5).setScale(0,BigDecimal.ROUND_HALF_UP).longValue());
 //        System.out.println(new BigDecimal(0.5).setScale(0,BigDecimal.ROUND_HALF_UP));
-        System.out.println("totalPrice  " + totalPrice(100L,5L));
-        System.out.println("totalRoundingPrice  " + totalRoundingPrice(100L,5L));
+        System.out.println("totalPrice  " + totalPrice(100L,5));
+        System.out.println("totalRoundingPrice  " + totalRoundingPrice(100L,5));
     }
 }
